requirements.txt
streamlit 
langchain
langchain_community
langchain_core
python-dotenv
langchain-huggingface
langchain-qdrant
langchain-ollama
unstructured
onnx==1.16.1
qdrant-client
python-pptx
python-docx
